breaking: updated niri, moved to lix, disabled chromium and removed steam.
Niri was slowing down my eval times, as it was pulling from a different cache.
so it the niri package now relies on nixpkgs version of niri. xwayland now works
on niri. Moved to the Lix version of nix, to speed up eval times. disabled
chromium for now as it was taking too long download the widevine.